Update project configuration
This commit is the pre-stage of changes before the remaining commits are added.
The important project configuration changes are stated below.

iOS:
- The version number follows semantic versioning by including PATCH.
- The build number has been reset and is no longer incremented.
- The phone can no longer view the app in landscape mode.

macOS:
- The version number follows semantic versioning by including PATCH.

The semantic versioning was always planned to be enforced, but as the app is still in early development, there was no reason to change it.

The build number does not reflect what version truly matters to us as we'll only be focusing on the future. It's useful to know the past, but
diving back into what build number we're on is not very useful. If semantic versioning on iOS supports labels (e.g. "-rc"), there will be little
to no reason to stick with build numbers.

The app can no longer be viewed in landscape as this caused view layout issues. In the future, I plan one bringing support back for landscape
mode (hopefully before the app is submitted to the App Store). Until then, it's easier to develop not having to worry about how the view looks
when in a transformed state.