Move tab management to tab navigation view
Prior to this commit, tab views were stored in the `Tabs/` directory. Some tabs stored the view, while some tabs acted like redirects.
This commit removes the `Tabs/` directory in favor of pointing to the view. Having a view redirect to another view was not very useful.
The only use case was for the developer's consistency, by adding another layer to redirect to. This allowed the developer to easily
glance at what were available, but this could be argued as redundant, as the developer could do the same in `TabNavigationView.swift`.
It's also easier to manage state, as you can be explicit about if the tab should appear based on the state of `CurrentUser`.