Move ALMetadata and HTTPNetworkTransportPreflightDelegate to their own scope
ALMetadata has been moved to `Metadata/AniList.swift` with the `ALMetadata` declared there.
The HTTPNetworkTransportPreflightDelegate protocol takes place in the same `Networking/ApolloNetwork.swift`
file, but is applied as an extension. This helps organize what's being inherited and from where.

This commit also replaces `@AppStorage(SettingsKeys.accessToken)` for `UserDefaults.standard.string(forKey:)`
because the AppStorage property would never be re-invoked, even if it was changed somewhere else (I think).
As the access token is retrieved at runtime, there shouldn't be worry about the token not being updated in the
`networkTransport` method.