Add ActivityFeed & remove UserActivity GraphQL queries
`UserView` will be removed in a later commit, so this GraphQL query is being removed in the meantime.
At the time, there plans to be two queries related to activity: the Global/Following feed, and a user feed.
The two queries will be called `ActivityFeed` and `UserActivity`. There was plan to have one GraphQL query
to do both Following/Global & user activity, but the AniList API returns an empty array when `Page.activities`'s
`userId` argument is `null`. This seems to be an issue, considering it does not make sense to not throw an error,
or factor in the possibility the requester meant any user. A GitHub issue has been opened to address this issue:
https://github.com/AniList/ApiV2-GraphQL-Docs/issues/106

Downloading the repository from this commit will likely result is compilation errors. This should only be used as
a reference to previous changes.