Remove user views
At the moment, priority has been set on creating the activity view. The user view is very incomplete and a long process,
so the views have been removed for the time being. Instead, the history will live on in the version control if I ever need
to access it again.