Remove logged in user modifier
Previously, we'd use `.navigationBarItems` to display the current user logged in (or a sign in button). This method has been deprecated,
and the alternative is to use `.toolbar`. However, `.toolbar` cannot display both the name and icon at the same time. Not having both can
be confusing for the user, so rather than sticking with the deprecated method, I've decided to remove it. Instead, operations requiring user
authentication will notify the user.`