Ignore error instead of printing it.
The cause of the error is most likely going to be 401 Unauthorized. This occurs when
fetching for a user when the user has not logged in. In the future, a light alert view will
appear to inform the user to log in. The soon-to-be workaround will be to use the native
`Alert` view, obstructing the user's view. However, this won't always be displayed to the
user (in cases like this).