Complete list editor + add multi-user support
The list editor is a view presented as a sheet to edit the entry on their list.
The user can edit all fields you can on https://anilist.co/, but features smart
fields to automatically fill in based on the value of other fields. For example:

- If the user has the anime/manga set to planning and increments the episodes/chapters
field, the status field will be set to watching/reading.
- If the user hasn't watched any episodes or read any chapters and increments it, the start
date will be set to the current date. If the user finishes it, the finish date is set to the current
date as well.

The `priority` field from the GraphQL API is displayed and incremental in the view. However,
it's not available on the website.

Multi-user support has also been added. It was added in the middle of the development as
swapping is much easier than logging out, logging back in, and authenticating. All users will
be fetched and loaded when the app launches, and Amincapp will acknowledge when the
current user has changed.