Complete text activity view & remove message activity view
This commit finishes the `ActivityTextKindView` view so it displays the user, timestamp, & content views, features the like & replies buttons, and context menu configuration. Although it's complete, some features haven't been implemented. For example, the replies button is not interactable (reason: https://github.com/LiteLT/Amincapp-Apple/issues/2).

References to message activities (not the same as text activities) have been removed as it's not in the global or following feed.

The `RichView` file (which was responsible for removing HTML content) has been removed. In the future, an HTML parser will be used to parse HTML & AniList-flavored markdown to native SwiftUI. The second alternative is to use a WekKit web view, but that brings overhead in the process.