Add Sidebar view
The sidebar serves as a replacement to the standard TabView we've been using. This will allow us to house more navigation views to destinations the user probably cares about, but it brings the downside of needing to pull up the sidebar to see it. On an iPad, this is fine since the user can always access the sidebar. On an iPhone, however, unless the screen is in landscape mode, the user must navigate all the way back to the first view to access the sidebar. The user could hold down on the back button to see a list of previous views they were at, but it would be nicer if the sidebar was always accessible.

The second issue with sidebars is how they perform when hidden. ContentView.swift explains the issue with the way SwiftUI handles presenting views.