Remove AniList distinction
This commit removes folders, files, and metadata that classifies AniList under its own group. For example, the `Shared/GraphQL` folder had two folders specific for AniList: `./AniList` and `./API/AniList`.

The reason for this removal is the app was originally drafted with the idea of using multiple services in one app (AniList, MyAnimeList, Kitsu, AniDB, etc). However, at the time of writing this, it may be more preferably and organizable to split native support for those list sites into their own projects. While this may encourage a project name change, it's unlikely to happen unless this project reaches a peak where it looks like a possibility with lots of reward to it.