Split Apollo interceptors and create nil removing body creator
The interceptors used to be located in the same file as GraphQLNetwork. Now, they're located under the `Interceptors` folder under the `Networking` folder.

The NilAwareRequestBodyCreator struct solves a major problem with the way AniList expects values and how Apollo sends them. To explain, there are three kind of variables that can be sent to a server. For demonstration, the variable is "five" with the value 5.
  - A present variable (`{ "five": 5 }`)
  - A null variable (`{ "five": null }`)
  - No variable (`{}`)

AniList expects the last one for most operations. In other cases, it avoids returning an error, but still does not handle the request the way we want it to. Apollo would use the second kind, which resulted in operations using multiple variables in fields using multiple variables that may or may not be present failing. A common example of this would be the `ToggleFavourite` mutation.

The downside to this solution is maintaining code that may change with Apollo versions, since the `requestBody` is just a copy-paste of Apollo's `ApolloRequestBodyCreator` that strips nil values. However, not moving versions and changing when necessary is simple.