Create media summary property views & simplify header view
All the property views used to be in the same file (MediaSummaryPropertiesView.swift). The views have now been broken up with the same structure as Shared/Features/Media/Sections/Properties.

The previous cover image took up 180x270 in dimensions, which was about 1/2 the screen horizontally. While this looked fine, it was too large to be decently viewed. It required a border to not look weird and ran into spacing issues for other views placed horizontally, such as the Share button with the Favorite and Edit buttons. This was especially troublesome for users with smaller phones, such as the iPhone 12 Mini and older phones. The cover image now takes up 117x176 in dimensions, which is far more maintainable, removes the border, and can fit all the buttons with extra space.