Create CurrentUserStore and update list views
CurrentUserStore acts as the control center for managing a user's accounts. Instead of storing the JWT access token from AniList in the user's default database, it instead stores them in the app's Keychain database. It's also going to clean up a few flaws and bad API design practices in the previous one. Although users are currently limited to one at most, this plans to change in the future.